linux/generic: (APP-3006, PR #247) force dev_coredump on
Sadly, KConfig mechanism for enabling CONFIG_DEV_COREDUMP is for
an (in-tree) driver to do CONFIG_WANT_DEV_COREDUMP.

But because the morse driver is built separately, it has no
way to demand this, and there's no way to manually specify
CONFIG_DEV_COREDUMP (or CONFIG_WANT_DEV_COREDUMP for that
matter). Therefore we hack the kconfig files.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn