boards: (APP-3496, PR #310) don't include dpp key recovery on artini
To achieve this, we repurpose dppd_diffconfig to be a
'include dpp qrcode support' thing, which implies
dpp-key-recovery (to recovery/generate the private key for
the STA side) and dppd (for the AP side).

Note that netifd-morse is still responsible for generating
the qrcode, as there's nasty 'look at the mac address
on initial module load' hacked in, but it now only
does this if the dpp private key exists.

Approved-by: Evan Benn
Approved-by: Milad Mohtashamirad