ramips: dts: (APP-3332, PR #314) add colour based aliases
This removes the extra functionality based aliases from the ekh03
(which are no longer used) and adds colour based aliases, which
make the diag_override.sh script clearer when we're using
multiple LEDs to construct a colour.

Note that we leave the old aliases in so that the board
will work without our custom LED package (morse-led-rgb).

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn