boards: add virtual wire configuration fragment
The virtual wire package allows for devices to be configured as an
access point or station with a jumper or switch. Assosciation is
handled by dpp push button, and channel selection by DCS on the
largest bandwidth for a region configured at device production.

Include the package as a configuration option.